fix: remove uncessary socket related code
